feat: add createVaultFolder functionality and refactor related components for improved folder management
